backend/requirements.txt
Django==5.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.4
openpyxl==3.1.2
groq==0.9.0
gunicorn==21.2.0
httpx==0.27.0
